Conflicts:
	java/com.sap.sailing.domain.common/src/com/sap/sailing/domain/common/impl/PolarSheetGenerationSettingsImpl.java
	java/com.sap.sailing.domain.test/src/com/sap/sailing/domain/test/PolarSheetGenerationTest.java
	java/com.sap.sailing.domain/META-INF/MANIFEST.MF
	java/com.sap.sailing.domain/src/com/sap/sailing/domain/polarsheets/PerRaceAndCompetitorPolarSheetGenerationWorker.java
	java/com.sap.sailing.domain/src/com/sap/sailing/domain/polarsheets/PolarFix.java
	java/com.sap.sailing.gwt.ui/META-INF/MANIFEST.MF
	java/com.sap.sailing.gwt.ui/SailingGWT.launch
	java/com.sap.sailing.gwt.ui/src/main/java/com/sap/sailing/gwt/ui/client/SailingService.java
	java/com.sap.sailing.gwt.ui/src/main/java/com/sap/sailing/gwt/ui/client/SailingServiceAsync.java
	java/com.sap.sailing.gwt.ui/src/main/java/com/sap/sailing/gwt/ui/client/StringMessages.java
	java/com.sap.sailing.gwt.ui/src/main/java/com/sap/sailing/gwt/ui/client/StringMessages.properties
	java/com.sap.sailing.gwt.ui/src/main/java/com/sap/sailing/gwt/ui/polarsheets/AngleOverDataSizeHistogramPanel.java
	java/com.sap.sailing.gwt.ui/src/main/java/com/sap/sailing/gwt/ui/polarsheets/PolarSheetsChartPanel.java
	java/com.sap.sailing.gwt.ui/src/main/java/com/sap/sailing/gwt/ui/polarsheets/PolarSheetsPanel.java
	java/com.sap.sailing.gwt.ui/src/main/java/com/sap/sailing/gwt/ui/server/SailingServiceImpl.java
	java/com.sap.sailing.mongodb.test/META-INF/MANIFEST.MF
	java/com.sap.sailing.polars.test/.classpath
	java/com.sap.sailing.polars.test/META-INF/MANIFEST.MF
	java/com.sap.sailing.polars.test/src/com/sap/sailing/polars/caching/test/PolarFixCacheRaceIntervalTest.java
	java/com.sap.sailing.polars.test/src/com/sap/sailing/polars/mining/test/PolarDataMinerTest.java
	java/com.sap.sailing.polars.test/src/com/sap/sailing/polars/test/PolarSheetAnalyzerTest.java
	java/com.sap.sailing.polars/.classpath
	java/com.sap.sailing.polars/src/com/sap/sailing/polars/PolarDataService.java
	java/com.sap.sailing.polars/src/com/sap/sailing/polars/aggregation/PolarFixAggregator.java
	java/com.sap.sailing.polars/src/com/sap/sailing/polars/aggregation/PolarFixRaceInterval.java
	java/com.sap.sailing.polars/src/com/sap/sailing/polars/aggregation/SimplePolarFixRaceInterval.java
	java/com.sap.sailing.polars/src/com/sap/sailing/polars/analysis/PolarSheetAnalyzer.java
	java/com.sap.sailing.polars/src/com/sap/sailing/polars/analysis/impl/PolarSheetAnalyzerImpl.java
	java/com.sap.sailing.polars/src/com/sap/sailing/polars/caching/PolarFixCacheRaceInterval.java
	java/com.sap.sailing.polars/src/com/sap/sailing/polars/clusters/SpeedClusterGroup.java
	java/com.sap.sailing.polars/src/com/sap/sailing/polars/impl/PolarDataServiceImpl.java
	java/com.sap.sailing.polars/src/com/sap/sailing/polars/mining/GPSFixMovingWithPolarContext.java
	java/com.sap.sailing.polars/src/com/sap/sailing/polars/mining/IncrementalRegressionProcessor.java
	java/com.sap.sailing.polars/src/com/sap/sailing/polars/mining/PolarDataMiner.java
	java/com.sap.sailing.polars/src/com/sap/sailing/polars/mining/PolarFixFilterCriteria.java
	java/com.sap.sailing.polars/src/com/sap/sailing/polars/mining/SpeedClusterGroupFromWindSteppingCreator.java
	java/com.sap.sailing.polars/src/com/sap/sailing/polars/regression/BoatSpeedEstimator.java
	java/com.sap.sailing.polars/src/com/sap/sailing/polars/regression/IncrementalLinearRegressionProcessor.java
	java/com.sap.sailing.polars/src/com/sap/sailing/polars/regression/impl/IncrementalLeastSquaresProcessor.java
	java/com.sap.sailing.server.gateway/META-INF/MANIFEST.MF
	java/com.sap.sailing.server.gateway/src/com/sap/sailing/server/gateway/jaxrs/api/PolarResource.java
	java/com.sap.sailing.server.replication.test/META-INF/MANIFEST.MF
	java/com.sap.sailing.server/META-INF/MANIFEST.MF
	java/com.sap.sailing.server/SailingServer (No Proxy).launch
	java/com.sap.sailing.server/SailingServer (No Proxy, Cached MTB).launch
	java/com.sap.sailing.server/SailingServer (No Proxy, Jetty on 8889).launch
	java/com.sap.sailing.server/SailingServer (No Proxy, Jetty on 8889, Cached MTB).launch
	java/com.sap.sailing.server/SailingServer (No Proxy, Remote Debug SAP VM).launch
	java/com.sap.sailing.server/SailingServer (No Proxy, winddbTest).launch
	java/com.sap.sailing.server/SailingServer (Proxy).launch
	java/com.sap.sailing.server/SailingServer (Proxy, Jetty on 8889).launch
	java/com.sap.sailing.server/SailingServer (Proxy, Remote Debug SAP VM).launch
	java/com.sap.sailing.server/SailingServer (Proxy, SwissTiming Live Simulation).launch
	java/com.sap.sailing.server/SailingServer (Proxy, winddbTest).launch
	java/com.sap.sailing.server/src/com/sap/sailing/server/RacingEventService.java
	java/com.sap.sailing.server/src/com/sap/sailing/server/impl/RacingEventServiceImpl.java
	java/com.sap.sailing.targetplatform/definitions/race-analysis-p2-remote.target
	java/com.sap.sse.datamining.shared/src/com/sap/sse/datamining/shared/impl/NestingCompoundGroupKey.java
	java/com.sap.sse.datamining/src/com/sap/sse/datamining/factories/GroupKeyFactory.java
	java/com.sap.sse.datamining/src/com/sap/sse/datamining/impl/components/AbstractParallelMultiDimensionalNestingGroupingProcessor.java
	java/pom.xml

+++ java/com.sap.sailing.domain.common/src/com/sap/sailing/domain/common/impl/WindSteppingImpl.java
@@ -11,9 +11,9 @@ public class WindSteppingImpl implements WindStepping {
     protected WindSteppingImpl() {};
 
     private static final long serialVersionUID = 2215693490331489508L;
-    protected Integer[] levels;
+    protected Double[] levels;
 
-    public WindSteppingImpl(Integer[] levels) {
+    public WindSteppingImpl(Double[] levels) {
         this.levels = levels;
     }
     
@@ -24,7 +24,7 @@ public class WindSteppingImpl implements WindStepping {
     @Override
     public int getLevelIndexForValue(double speed) {
         for (int i = 0; i < levels.length - 1; i++) {
-            if (speed < levels[i] + ((levels[i+1] - levels[i]) / 2)) {
+            if (speed < levels[i] + ((levels[i+1] - levels[i]) / 2.)) {
                 return i;
             }  
         }
@@ -32,14 +32,14 @@ public class WindSteppingImpl implements WindStepping {
     }
 
     @Override
-    public Integer[] getRawStepping() {
+    public Double[] getRawStepping() {
         return levels;
     }
 
     @Override
-    public int getSteppedValueForValue(double speed) {
+    public Double getSteppedValueForValue(double speed) {
         int levelIndex = getLevelIndexForValue(speed);
-        int result = - 1;
+        Double result = -1.0;
         if (levelIndex >= 0) {
             result = levels[levelIndex];
         }
